Read and search for a tag within a spreadsheet
Added functionality to read an excel spreadsheet and find a cell by a step number. This will be used to create a dictionary of the valves in that step as keys and their state as values. This dictionary will then be used to compare the actual valve states against the screenshot that has been converted to text from the live system.